+++ src/utils/stopwatch.ts
@@ -7,8 +7,12 @@ const NS_PER_SEC = 1e9;
  * measure multiple Times is also possible with SectionStart and SectionElapsedMs
  */
 export class StopWatch {
-    private readonly _time: number | [number, number];
+    private _time: number | [number, number];
     private readonly _multipleTimes: {[key: string]: number | [number, number]};
+    private _measures: number[] = [];
+    private _isPause = false;
+    private _multipleMeasures: {[key: string]: number[]} = {};
+    private _multipleIsPause: {[key: string]: boolean} = {};
 
     /**
      * create a new StopWatch Instance at this Time the StopWatch was started
@@ -42,6 +46,23 @@ export class StopWatch {
      */
     public SectionStart(key: string): void {
         this._multipleTimes[key] = this.getTimestamp();
+        this._multipleMeasures[key] = [];
+        this._multipleIsPause[key] = false;
+    }
+
+    /**
+     * checks if a Section was paused
+     * @param key the key that indicates the Section
+     */
+    public IsSectionPause(key: string): boolean {
+        return this._multipleIsPause[key] === true;
+    }
+
+    /**
+     * if the current StopWatch in pause mode
+     */
+    get IsPause(): boolean {
+        return this._isPause === true;
     }
 
     /**
@@ -49,14 +70,66 @@ export class StopWatch {
      * @param key the key that indicates the Section
      */
     public SectionElapsedMs(key: string): number {
-        return this.getMultipleTimeDiff(key);
+        let tmp = this._multipleIsPause[key] ? 0 : this.getMultipleTimeDiff(key);
+        if (isNaN(tmp)) {
+            tmp = 0;
+        }
+        if (!this._multipleMeasures[key]) {
+            this._multipleMeasures[key] = [];
+        }
+        if (this._multipleMeasures[key].Any()) {
+            for (const t of this._multipleMeasures[key]) {
+                tmp += t;
+            }
+        }
+        return tmp;
+    }
+
+    /**
+     * same as Pause only for Sections
+     * @param key the key that indicates the Section
+     */
+    public SectionPause(key: string) {
+        this._multipleMeasures[key].Add(this.getMultipleTimeDiff(key));
+        this._multipleIsPause[key] = true;
+    }
+
+    /**
+     * same as Resume only for Sections
+     * @param key the key that indicates the Section
+     */
+    public SectionResume(key: string) {
+        this._multipleIsPause[key] = false;
+        this._multipleTimes[key] = this.getTimestamp();
+    }
+
+    /**
+     * stops the StopWatch from measure Time
+     */
+    public Pause() {
+        this._measures.Add(this.getTimeDiff());
+        this._isPause = true;
+    }
+
+    /**
+     * starts the StopWatch at the Point from the Last Pause
+     */
+    public Resume() {
+        this._time = this.getTimestamp();
+        this._isPause = false;
     }
 
     /**
      * gets the Elapsed Time in ms from the StopWatch
      */
     public ElapsedMs(): number {
-        return this.getTimeDiff();
+        let tmp = this._isPause ? 0 : this.getTimeDiff();
+        if (this._measures.Any()) {
+            for (const t of this._measures) {
+                tmp += t;
+            }
+        }
+        return tmp;
     }
 
     private getTimestamp(): number | [number, number] {